pyproject.toml
[project]
name = "room-type-classifier"
version = "0.1.0"
description = "Room type classifier experiments"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
]

[dependency-groups]
data = [
  "numpy==2.4.3",
  "pandas==2.3.3",
  "pillow==12.1.1",
  "scikit-learn==1.5.2",
  "torch==2.5.1",
  "torchvision==0.20.1",
]
efficientnet = [
  "numpy==2.4.3",
  "pandas==2.3.3",
  "pillow==12.1.1",
  "scikit-learn==1.5.2",
  "torch==2.5.1",
  "torchvision==0.20.1",
]
interpretability = [
  "grad-cam==1.5.5",
  "matplotlib==3.10.9",
  "opencv-python==4.11.0.86",
]
streamlit = [
  "pandas==2.3.3",
  "pillow==12.1.1",
  "streamlit==1.55.0",
]
yolo = [
  "huggingface-hub==1.12.0",
  "opencv-python==4.11.0.86",
  "pandas==2.3.3",
  "pillow==12.1.1",
  "protobuf==6.33.6",
  "torch==2.5.1",
  "torchvision==0.20.1",
  "ultralytics==8.0.20",
  "ultralyticsplus==0.0.21",
]

[tool.uv]
managed = true
